# api/openapi.yaml
...
openapi: '3.0.0'
info:
  description: ''
  version: 0.0.0
  title: dgc API
host: localhost:9001
basePath: /
...
schemes:
  - http
paths:
...
  /batch_status:
    get:
      tags:
        - transactions
      operationId: get_batch_status
      parameters:
        - name: wait
          in: query
          description: How long the rest API should wait for transactions to be committed
          schema:
            type: integer
        - name: ids
          in: query
          required: true
          description: Comma separated list of batch ids we want to retrieve the status of
          schema:
            type: string

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#definitions/BatchStatuses'
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Invalid Id
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unexpected error occurred while submitting batches
...